get zql -> materialization working e2e
